Minor Scale fixes for enemies. (#137)
Checkout the verminator's hitboxes to see if i messed
anything up. As for the rest, should be fine

Oh and the main_menu.gd and score_line are changes
are just adding "_" into function parameters on declaration.
These warnings are doing my head in.

This isn't a big deal, right?