Launch sequence (#154)
* The First Commit!

* TODO,
find an alternative fit for the
currtent scene setup. shouldn't be that hard

* Cutscene data folder added

* Added Music

* Godot fucking up again
something wit the language server i believe

* Figers crossed

* Cutscene Node ver 1.0 commit
With this thing I'm aiming for "flexible" framework for creating "retro-styled cutsscenes"
"cutscene" node is to be provided with a cutscene title that corresponds to directories for media
"CutsceneData" folder. As for now, we have only one ready-to-go cutscene.
Setting up a cutscene node entails:
   1)Adding it to the tree
   2)Settign its string property called "scene_tile" (any naming i do is subject to chage)
   to the name of a cutscene you wish to launch (ex. "opening", upper/lower case don matter)
   3)calling launch_scene() function...
Note: Line 45, the thing calls launch_scene from _ready(). gotta erase it later

* fix merge issue

* fix verminator merge

---------

Co-authored-by: Stephan Auerhahn <stephan@stability.ai>